Trading Signal/requirements.txt
aiokafka==0.10.0
annotated-types==0.6.0
async-timeout==4.0.3
asyncio==3.4.3
packaging==23.2
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.1
typing_extensions==4.9.0
